+++ spec/dummy/app/models/complex_configuration.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class ComplexConfiguration < Configuration
-  attribute :suppliers, Supplier.to_array_type
-end